new file:   P2252.cpp 	new file:   "c++\346\226\207\344\273\266/GESP 7\347\272\247\345\244\207\350\200\203/GESP7-\351\200\237\346\237\245\346\211\213\345\206\214-v1.0.pdf" 	modified:   "code\346\226\207\344\273\266/text.cpp" 	new file:   "code\346\226\207\344\273\266/text2.cpp" 	new file:   "\346\250\241\346\235\277\345\217\212\351\207\215\350\246\201\344\273\243\347\240\201/\346\240\221\347\212\266\346\225\260\347\273\204(\345\214\272\351\227\264\344\277\256\346\224\271\345\215\225\347\202\271\346\237\245\350\257\242).cpp" 	new file:   "\346\250\241\346\235\277\345\217\212\351\207\215\350\246\201\344\273\243\347\240\201/\346\240\221\347\212\266\346\225\260\347\273\204(\345\215\225\347\202\271\344\277\256\346\224\271\345\214\272\351\227\264\346\261\202\345\222\214).cpp"

+++ P2252.cpp
@@ -3,10 +3,14 @@ using namespace std;
 #define ll long long
 #define el '\n'
 const ll N=5e6;
-const double g=(1.0+sqrt(5.0))/2.0;
+const long double g=(1.0+sqrtl(5.0))/2.0;
 
 void solve(){
     ll a,b;cin>>a>>b;
+    // if(a==433494437&&b==701408733){
+    //     cout<<1;
+    //     return;
+    // }
     if(a>b) swap(a,b); 
     ll k=b-a;
     cout<<(a==floor(g*(k*1.0))?"0":"1");
